% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plotting.R
\name{autoplot_methods}
\alias{autoplot_methods}
\alias{autoplot.tornado_df}
\alias{autoplot.TornadoExperiment}
\title{Autoplot methods for tornado plots}
\usage{
\method{autoplot}{tornado_df}(object, ..., facet = TRUE, x_scale = TRUE, y_scale = TRUE)

\method{autoplot}{TornadoExperiment}(
  object,
  ...,
  upper = "q0.99",
  lower = 0,
  scale_title = "Coverage",
  scale = NULL
)
}
\arguments{
\item{object}{A \linkS4class{TornadoExperiment} or \code{tornado_df} object.}

\item{...}{Used in the \linkS4class{TornadoExperiment} method to pass the
\code{facet}, \code{x_scale} and \code{y_scale} arguments to the \code{tornado_df} method.}

\item{facet}{A \code{logical(1)} or ggplot2 facet. If \code{TRUE}, a
\code{\link[=facet_tornado]{facet_tornado()}} is added. If \code{FALSE}, no facet will
be added. When a ggplot2 facet, the facet is added to the plot.}

\item{x_scale}{A \code{logical(1)} or ggplot2 x scale. If \code{TRUE}, an x scale is
added that attempts to avoid overlapping labels.}

\item{y_scale}{A \code{logical(1)} or ggplot2 y scale. If \code{TRUE}, a y scale is
added that marks regular intervals but only labels the number of features.}

\item{upper}{Limits to the colour scale. Can be one of the following:
\itemize{
\item A \code{numeric} to set absolute limits directly.
\item A \code{character} in the form of a number prefixed by \code{"p"} or \code{"q"} to set
the percentile or quantile respectively. For example, \code{"p99"} and \code{"q0.95"}
set the limit to the 99th percentile or 0.95th quantile respectively. The
limits as specified in the \code{lower} and \code{upper} arguments can be overridden
if the \code{scale} argument has non-default limits.
}}

\item{lower}{Limits to the colour scale. Can be one of the following:
\itemize{
\item A \code{numeric} to set absolute limits directly.
\item A \code{character} in the form of a number prefixed by \code{"p"} or \code{"q"} to set
the percentile or quantile respectively. For example, \code{"p99"} and \code{"q0.95"}
set the limit to the 99th percentile or 0.95th quantile respectively. The
limits as specified in the \code{lower} and \code{upper} arguments can be overridden
if the \code{scale} argument has non-default limits.
}}

\item{scale_title}{A \code{character(1)} with the title for the colourbar. Can
be overridden when the \code{scale} argument has a non-default name.}

\item{scale}{A \link[ggplot2:scale_colour_continuous]{continuous fill scale} from
the ggplot2 package to use to colour the tornado. When \code{NULL} (default),
the scale will be retrieved from the option \code{"tornadoplot.default.scale"}.}
}
\value{
A ggplot object.
}
\description{
Autoplot methods for tornado plots
}
\examples{
NULL
}
