% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_tornado.R
\name{collapse_tornado}
\alias{collapse_tornado}
\title{Collapse tornado features}
\usage{
collapse_tornado(tornado, fun = colMeans, ...)
}
\arguments{
\item{tornado}{A \code{tornado_array} dataset.}

\item{fun}{A function to apply to a matrix build from every section of the
array.}

\item{...}{Extra arguments to pass to the \code{fun} function.}
}
\value{
A \code{data.frame} with \code{value}, \code{feature_set}, \code{position} and \code{sample}
columns.
}
\description{
Use a summarising function to collapse a tornado along the features.
}
\examples{
NULL
}
