% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plotting.R
\name{geom_tornado}
\alias{geom_tornado}
\title{Tornado Heatmap}
\usage{
geom_tornado(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  interpolate = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{A \code{tornado_df} data.frame, such as one that results from
processing a tornado array with \code{\link[=prep_tornado]{prep_tornado()}}.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{interpolate}{If \code{TRUE} interpolate linearly, if \code{FALSE}
(the default) don't interpolate.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
A \code{GeomTornado} ggproto object that can be added to a plot.
}
\description{
\code{geom_tornado()} serves as a plotting layer for tornado heatmaps.
}
\details{
The required aesthetics are synonymous with the column names of a
default \code{tornado_df}.
}
\section{Aesthetics}{

\code{geom_tornado} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
\item \strong{xmin}
\item \strong{xmax}
\item \strong{ymin}
\item \strong{ymax}
\item \strong{tornado}
}
}

\examples{
NULL
}
