% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_adex.R
\name{init_adex}
\alias{init_adex}
\title{Create a small ADEX dataset using SDTM sources datasets and settings}
\usage{
init_adex(data, settings)
}
\arguments{
\item{data}{named list of data}

\item{settings}{named list of settings}
}
\value{
returns list with ADEX data added to other data and settings
}
\description{
Create a small ADEX dataset using SDTM sources datasets and settings
}
\examples{

init_adex(data=params$data, settings=params$settings)
}
