% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_adsl.R
\name{init_adsl}
\alias{init_adsl}
\title{Create a small ADSL dataset using SDTM sources datasets and settings}
\usage{
init_adsl(data, settings)
}
\arguments{
\item{data}{named list of data}

\item{settings}{named list of settings}
}
\value{
returns list with ADSL data added to other data and settings
}
\description{
Create a small ADSL dataset using SDTM sources datasets and settings
}
\examples{

init_adsl(params$data, settings=params$settings)
}
