% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_enrolmap.R
\name{init_enrolmap}
\alias{init_enrolmap}
\title{Initialize Settings and Data for Enrollement map widget}
\usage{
init_enrolmap(
  data,
  settings,
  ref_arm = "Treatment",
  comp_arm = "Placebo",
  threshold = 1
)
}
\arguments{
\item{data}{named list of data}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings used for fct_tornadoplot
}
\description{
Initialize Settings and Data for Enrollement map widget
}
\examples{

init_tornado(params$data, settings=params$settings, ref_arm= "Treatment", comp_arm = "Placebo")
}
