% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_exposure.R
\name{init_exposure}
\alias{init_exposure}
\title{Initialize Settings and Input data for Exposure plot (Swimmer plot) widget}
\usage{
init_exposure(
  data,
  settings,
  ref_arm = "Treatment",
  comp_arm = "Placebo",
  threshold = 1
)
}
\arguments{
\item{data}{named list of data}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings used for fct_tornadoplot
}
\description{
Initialize Settings and Input data for Exposure plot (Swimmer plot) widget
}
\examples{

init_exposure(params$data, settings=params$settings, ref_arm= "Treatment", comp_arm = "Placebo", threshold = 1)
}
