% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_params.R
\name{init_params}
\alias{init_params}
\title{Initialize Settings and Data for the app, fill missing settings if not provided by user}
\usage{
init_params(data, settings)
}
\arguments{
\item{data}{named list of data}

\item{settings}{named list of settings}
}
\value{
returns list with data and settings used through the app. The data is loaded by the user as explained in the README
}
\description{
Initialize Settings and Data for the app, fill missing settings if not provided by user
}
\examples{
init_params(data = list(ex = sdtm_ex, dm = sdtm_dm, ds = sdtm_ds), settings = list(ex=list(id_col="SUBJID")))

}
