% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_tornado.R
\name{melt_tornado}
\alias{melt_tornado}
\title{Reshape tornado to long format}
\usage{
melt_tornado(tornado, assay_name = assayNames(tornado))
}
\arguments{
\item{tornado}{A \linkS4class{TornadoExperiment} object.}

\item{assay_name}{A \code{character(1)}: one of the assay names.}
}
\value{
A \code{data.frame}
}
\description{
Reshape tornado to long format
}
\examples{
NULL
}
