% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tornado.R
\name{mod_tornado_ui}
\alias{mod_tornado_ui}
\title{tornado UI Function}
\usage{
mod_tornado_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
