% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_tornado.R
\name{norm_tornado}
\alias{norm_tornado}
\title{Normalise samples}
\usage{
norm_tornado(
  tornado,
  scale,
  assay_name = assayNames(tornado),
  assay_to = assay_name
)
}
\arguments{
\item{tornado}{A \linkS4class{TornadoExperiment} object.}

\item{scale}{A \code{numeric} vector or length \code{ncol(tornado)} with normalisation
factors.}

\item{assay_name}{A \code{character(1)}: one of the assay names.}

\item{assay_to}{A \code{character(1)} for the target assay name for the normalised
data. Defaults to the \code{assay_name} assay, thereby replacing input data.}
}
\value{
A \linkS4class{TornadoExperiment} object.
}
\description{
Normalise the samples in a \linkS4class{TornadoExperiment} by dividing their
values by some scaling factor.
}
\examples{
tor <- dummy_tornado()
tor <- norm_tornado(tor, c(1, 2))
}
\seealso{
Other tornado utilities: 
\code{\link{flatten_features}()},
\code{\link{melt_tornado}()},
\code{\link{sort_tornado}()}
}
\concept{tornado utilities}
