% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_type}
\alias{norm_type}
\alias{norm_type_or_null}
\alias{norm_type,ANY,numeric-method}
\alias{norm_type,ANY,integer-method}
\title{Argument checking}
\usage{
norm_type_or_null(
  x,
  type,
  length = 1,
  allow_NAs = FALSE,
  arg = deparse(substitute(x))
)

norm_type(x, type, length = 1, allow_NAs = FALSE, arg = deparse(substitute(x)))

\S4method{norm_type}{ANY,numeric}(x, type, length = 1, allow_NAs = FALSE, arg = deparse(substitute(x)))

\S4method{norm_type}{ANY,integer}(x, type, length = 1, allow_NAs = FALSE, arg = deparse(substitute(x)))
}
\arguments{
\item{x}{An argument to check.}

\item{type}{A vector of the expected type, typically of length 0.}

\item{length}{An \code{integer(1)} or \code{NULL} for the expected length. When set to
\code{NULL}, length can be anything.}

\item{allow_NAs}{A \code{logical(1)}, when \code{TRUE}, allows the \code{x} argument to have
\code{NAs}. If \code{FALSE}, raises error when \code{x} contains \code{NA}s.}

\item{arg}{A \code{character(1)} for the argument being checked. This is used to
identify the argument in error messaging. Defaults to the symbol used for
\code{x}.}

\item{...}{Not currently used.}
}
\value{
It returns \code{x} if it satisfies the criteria, or \code{x} coerced to the
\code{type} class when appropriate.
}
\description{
These functions are for checking that arguments of parent functions are as
expected.
}
\examples{
# Rounding of `numeric` to `integer`
norm_type(1.26, integer())
}
