% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{octsox_peaks}
\alias{octsox_peaks}
\title{Sox2 and Oct4 binding sites}
\format{
A \linkS4class{GRangesList} with 3 elements, each having 500 genomic
locations of binging sites.
\describe{
\item{Sox2}{500 genomic loci where Sox2 binds}
\item{Oct4}{500 genomic loci where Oct4 binds}
\item{Both}{500 genomic loci where both Sox2 and Oct4 bind}
}
}
\source{
The files used to construct this data were downloaded as bed files from
\url{http://cistrome.org/db/#/} (accessed at 2021-08-18). The files can be
unambiguously found by entering the following numbers in the search bar:
"73468" for the Sox2 sample and "73466" for the Oct4 sample.
}
\description{
This dataset contains a sample of genomic binding sites of the Sox2 and Oct4
proteins in mouse embryonic stem cells.
}
\details{
The genomic loci were obtained by chromatin immunoprecipitation
followed by sequencing (ChIP-seq) with antibodies against the Sox2 and Oct4
proteins. The model system used were wildtype mouse embryonic stem cells.

The original data was deposited in the gene expression omnibus (GEO) under
accession
\href{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE74112}{GSE74112}.

The data were aligned against the "mm10" reference genome by Cistrome. Other
preprocessing steps are described at the
\href{http://cistrome.org/db/#/about}{Cistrome website}.

Finally, locations in standard chromosomes were reduced to unique sites and
categorised based on overlap with Sox2 binding sites and Oct4 binding sites
or both. From each of the three resultant categories, 500 locations were
randomly sampled without replacement.
}
\references{
Liu, Ziying \emph{et al}. Catalytic-Independent Functions of PARP-1 Determine Sox2
Pioneer Activity at Intractable Genomic Loci (2017) Molecular Cell \emph{65} \strong{4}
p589-603 https://doi.org/10.1016/j.molcel.2017.01.017
}
\keyword{datasets}
