% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_tornado.R
\name{prep_tornado}
\alias{prep_tornado}
\alias{tornado_df}
\title{Prepare tornado array for plotting}
\usage{
prep_tornado(
  tornado,
  assay_name = assayNames(tornado),
  upper = "q0.99",
  lower = 0,
  scale_title = "Coverage",
  scale = NULL
)
}
\arguments{
\item{tornado}{A \linkS4class{TornadoExperiment} object.}

\item{assay_name}{A \code{character(1)}: one of the assay names retrieved by
\code{assayNames(tornado)}.}

\item{upper, lower}{Limits to the colour scale. Can be one of the following:
\itemize{
\item A \code{numeric} to set absolute limits directly.
\item A \code{character} in the form of a number prefixed by \code{"p"} or \code{"q"} to set
the percentile or quantile respectively. For example, \code{"p99"} and \code{"q0.95"}
set the limit to the 99th percentile or 0.95th quantile respectively. The
limits as specified in the \code{lower} and \code{upper} arguments can be overridden
if the \code{scale} argument has non-default limits.
}}

\item{scale_title}{A \code{character(1)} with the title for the colourbar. Can
be overridden when the \code{scale} argument has a non-default name.}

\item{scale}{A \link[ggplot2:scale_colour_continuous]{continuous fill scale} from
the ggplot2 package to use to colour the tornado. When \code{NULL} (default),
the scale will be retrieved from the option \code{"tornadoplot.default.scale"}.}
}
\value{
A \code{tornado_df}, \code{data.frame} class object with a modified scale in
the \code{"scale"} attribute.
}
\description{
Converts a tornado array in a data.frame that is ready to be used with
the \code{geom_tornado()} layer for the ggplot2 library.
}
\details{
The typical way of plotting a raster in ggplot2 is with the
\code{\link[ggplot2:geom_tile]{geom_raster()}} function. However, this requires
you to melt the raster first only to later reconstitute it, which can be
rather inefficient. To circumvent this inconvenience, we pre-colour the
data here with a ggplot2 scale.
}
\examples{
NULL
}
