% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{BigWigFile}
\alias{BigWigFileList}
\alias{path}
\alias{metadata}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{BiocGenerics}{\code{\link[BiocGenerics]{path}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{rtracklayer}{\code{\link[rtracklayer]{BigWigFile}}, \code{\link[rtracklayer:BigWigFile]{BigWigFileList}}}

  \item{S4Vectors}{\code{\link[S4Vectors:Annotated-class]{metadata}}}
}}

