% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_tornado.R
\name{sort_tornado}
\alias{sort_tornado}
\title{Reorder features of a tornado}
\usage{
sort_tornado(
  tornado,
  assay_name = assayNames(tornado),
  bin_subset = NULL,
  bin_weights = NULL,
  sample_subset = NULL,
  decreasing = TRUE
)
}
\arguments{
\item{tornado}{A \linkS4class{TornadoExperiment} object.}

\item{assay_name}{A \code{character(1)}: one of the assay names.}

\item{bin_subset, bin_weights}{Define either \code{bin_subset} or \code{bin_weights} (or
neither), but not both. \code{bin_subset} should be a \code{logical} or \code{numeric}
vector with subsetting indices for the number of bins. \code{bin_weights} should
be a \code{numeric} vector of weights, the same length as there are bins. If
both \code{bin_subset} and \code{bin_weights} are \code{NULL} (default), the weights are
constructed with the density function of the Laplace distribution wherein
the location is centred at the middle of the bins and the scale parameter
is 10\% of the number of bins.}

\item{sample_subset}{A \code{character}, \code{logical} or \code{integer} vector for
subsetting the samples.}

\item{decreasing}{A \code{logical(1)}: sort from high to low (\code{TRUE}) or from low
to high (\code{FALSE})?}
}
\value{
The \code{tornado} dataset with the features reordered.
}
\description{
Reorder features of a tornado
}
\examples{
NULL
}
