% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_plotting.R
\name{tornado_plot}
\alias{tornado_plot}
\title{Make a tornado plot.}
\usage{
tornado_plot(..., sort = TRUE)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=build_tornado]{build_tornado}}
  \describe{
    \item{\code{features}}{A \linkS4class{GRanges} or \linkS4class{GRangesList} object
containing genomic loci of interest.}
    \item{\code{data}}{The data source to get coverages. Either a \linkS4class{GRanges},
\linkS4class{GRangesList}, \linkS4class{TabixFile},
\linkS4class{BigWigFile}, \linkS4class{BigWigFileList} object or
\code{character} path to \code{.bw} files.}
    \item{\code{width}}{An \code{integer(1)} in basepairs with respectively a
common feature width to centre features in, a size of bins to summarise
coverage in or the number of bins to summarise coverage in. Only two of
these need to be defined.}
    \item{\code{binwidth}}{An \code{integer(1)} in basepairs with respectively a
common feature width to centre features in, a size of bins to summarise
coverage in or the number of bins to summarise coverage in. Only two of
these need to be defined.}
    \item{\code{nbin}}{An \code{integer(1)} in basepairs with respectively a
common feature width to centre features in, a size of bins to summarise
coverage in or the number of bins to summarise coverage in. Only two of
these need to be defined.}
    \item{\code{pad_value}}{A \code{numeric(1)} to use for padding when
\code{\link[GenomeInfoDb:seqinfo]{seqlenghts(features)}} are unknown or
features exceed known sequence lengths.}
  }}

\item{sort}{A \code{logical(1)}. Should the tornado be sorted from high to low?}
}
\value{
A \linkS4class{TornadoExperiment} object with a \code{ggplot} object in
the \code{\link[S4Vectors:Annotated-class]{metadata()}} annotation.
}
\description{
The \code{tornado_plot()} function exists to easily jot down a tornado plot. It
will print the plot and return the data used for plotting with the plot
itself in the metadata. This function performs the tasks of retrieving and
aggregating data from its source with \code{\link[=build_tornado]{build_tornado()}}, optionally sorting
the result with \code{\link{sort_tornado}} and finally plotting.
}
\examples{
feats <- dummy_features()
file  <- dummy_granges_data()

tornado_plot(feats, file, width = 3000)
}
