% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TornadoApp.R
\name{TornadoApp}
\alias{TornadoApp}
\title{Tornado App}
\usage{
TornadoApp(dfAE = NULL, dfDemog = NULL, settings = NULL, runNow = TRUE)
}
\arguments{
\item{dfAE}{AE Data}

\item{dfDemog}{demog data}

\item{settings}{safetyGraphics settings}
}
\description{
Tornado App
}
