% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornadoplot-package.R
\docType{package}
\name{tornadoplot-package}
\alias{tornadoplot}
\alias{tornadoplot-package}
\title{tornadoplot: Tornado Plot}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Create plots of feature-centered aggregated genomic coverage tracks as heatmaps.
}
\author{
\strong{Maintainer}: Teun van den Brand \email{t.vd.brand@nki.nl} (\href{https://orcid.org/0000-0002-9335-7468}{ORCID})

}
\keyword{internal}
