% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tornadoplot_table.R
\name{tornadoplot_wtable}
\alias{tornadoplot_wtable}
\title{Create a tornado plot}
\usage{
tornadoplot_wtable(data, settings, groupvar = "None", ref_arm, comp_arm)
}
\arguments{
\item{data}{A dataframe list}

\item{settings}{Named list of settings}
}
\value{
a ggplot object render in the shinyapp
}
\description{
Create a tornado plot
}
\examples{
settings<-list(
  stratification_col="AEBODSYS",
  group_col="ARM",
  severity_col = "AESEV",
  serious_col = "AESER",
  reference_group="Placebo",
  comparison_group="Xanomeline High Dose",
  id_col="USUBJID"
)
tornadoplot(data, settings)

}
