% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_tornado.R
\name{melt_tornado}
\alias{melt_tornado}
\title{Reshape tornado to long format}
\usage{
melt_tornado(tornado, assay_name = assayNames(tornado))
}
\arguments{
\item{tornado}{A \linkS4class{TornadoExperiment} object.}

\item{assay_name}{A \code{character(1)}: one of the assay names.}
}
\value{
A \code{data.frame} with \code{prod(dim(tornado))} rows.
}
\description{
The \code{melt_tornado} emulates for the \linkS4class{TornadoExperiment} class
what the \code{melt()} function does in the \pkg{reshape2} package. It converts
the data to long format, which is the format of choice in the \pkg{tidyverse}
family of packages.
}
\note{
For plotting purposes, the \code{\link[=prep_tornado]{prep_tornado()}} function
does a more efficient job than \code{melt_tornado()}.
}
\examples{
# Melt a tornado
tor <- dummy_tornado()
# Reshaping to long format
df  <- melt_tornado(tor)

# Building a tornado plot from scratch
require(ggplot2)
ggplot(df, aes(position, feature_nr)) +
  geom_raster(aes(fill = value)) +
  facet_grid(feature_set ~ sample_name)
}
\seealso{
Other tornado utilities: 
\code{\link{flatten_features}()},
\code{\link{norm_tornado}()},
\code{\link{sort_tornado}()}
}
\concept{tornado utilities}
