% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{testdata}
\alias{testdata}
\alias{dummy_features}
\alias{dummy_granges_data}
\alias{dummy_tabix}
\alias{dummy_bigwig}
\title{Generation of test data}
\usage{
dummy_features()

dummy_granges_data()

dummy_tabix(pattern = "file", dir = tempdir())

dummy_bigwig(pattern = "file", dir = tempdir())
}
\arguments{
\item{pattern}{A \code{character(1)} giving an intial part of a file name.}

\item{dir}{A \code{character(1)} giving a directory name where to create the test
file.}
}
\value{
The return values differ per function. \describe{
\item{\code{dummy_features}}{A \code{GRanges} object of length 2.}
\item{\code{dummy_granges_data}}{A \code{GRanges} object of length 20.}
\item{\code{dummy_bigwig}}{A \code{BigWigFile} object.}
\item{\code{dummy_tabix}}{A \code{TabixFile} object with \code{"A"} and \code{"B"} in the 4th
column.}
}
}
\description{
For the purposes of examples and testing the package, tornadoplot includes
some functions that make up some dummy data that isn't too large.
}
\details{
All the dummy data generators make their data around the genomic
ranges returned by \code{dummy_features()}.

The \code{dummy_bigwig()} and \code{dummy_tabix()} files create files on your system.
As their usage is intended as temporary files, don't forget to \code{unlink()}
their paths and indices after they are no longer needed.
}
\examples{
# All the functions run without arguments
feats <- dummy_features()
gr    <- dummy_granges_data()
bw    <- dummy_bigwig()
tbx   <- dummy_tabix()

# Don't forget to clean up the temporary files when you're done
unlink(c(path(bw), path(tbx), Rsamtools::index(tbx)))
}
